% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_WPOACM.R
\name{run_WPROACM}
\alias{run_WPROACM}
\title{Run WPROACM}
\usage{
run_WPROACM()
}
\description{
Runs the WPROACM shiny application, a GUI for the statnet
suite of network analysis packages.
}
\examples{
\dontrun{
run_WPROACM()
}


}
\concept{networks GUI shiny ergm}
\keyword{datagen}
\keyword{graphs}
\keyword{models}
