% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPROACM-package.R
\docType{package}
\name{WPROACM}
\alias{WPROACM-package}
\alias{WPROACM}
\title{WPROACM: A graphical user interface for network analysis.}
\description{
The WPROACM package makes it easy to run the graphical user interface
for statistical packages.
}
\author{
\strong{Maintainer}: Mark S. Handcock \email{handcock@stat.ucla.edu}

Other contributors:
\itemize{
  \item Bart Blackburn \email{bartblackburn@ucla.edu} [contributor]
}

}
