% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run WPROACM}
\usage{
run()
}
\description{
Runs the WPROACM shiny application, a GUI for the statnet
suite of network analysis packages.
}
\examples{
\dontrun{
run()
}


}
\concept{networks GUI shiny ergm}
\keyword{datagen}
\keyword{graphs}
\keyword{models}
